Update email handler to include email read capabilities. Rename to email sender
