+++ src/Plugins/BotSharp.Plugin.EmailHandler/Providers/IEmailReader.cs
@@ -1,11 +1,11 @@
-using BotSharp.Plugin.EmailReader.Models;
+using BotSharp.Plugin.EmailHandler.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BotSharp.Plugin.EmailReader.Providers
+namespace BotSharp.Plugin.EmailHandler.Providers
 {
     public interface IEmailReader
     {
