+++ src/Plugins/BotSharp.Plugin.EmailHandler/Settings/EmailSenderSettings.cs
@@ -1,6 +1,6 @@
 namespace BotSharp.Plugin.EmailHandler.Settings;
 
-public class EmailHandlerSettings
+public class EmailSenderSettings
 {
     public string EmailAddress { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
