+++ src/Plugins/BotSharp.Plugin.EmailHandler/data/agents/6745151e-6d46-4a02-8de4-1c4f21c7da95/functions/handle_email_sender.json
@@ -1,5 +1,5 @@
 {
-  "name": "handle_email_request",
+  "name": "handle_email_sender",
   "description": "If the user requests to send an email with or without attachments or files, you need to capture the email content and the recipient email address. If the user explicitly enter email subject use the same if not intelligently capture the email subject from the content. Then call this function to send out email.",
   "parameters": {
     "type": "object",
