--- src/Plugins/BotSharp.Plugin.EmailHandler/data/agents/6745151e-6d46-4a02-8de4-1c4f21c7da95/templates/handle_email_request.fn.liquid
@@ -1,2 +1,0 @@
-Please call handle_email_request if user wants to send email.
-** Please take a look at the conversation and decide whether user wants to send email with files/attachments/images or not.
