--- src/Plugins/BotSharp.Plugin.EmailReader/BotSharp.Plugin.EmailReader.csproj
@@ -1,37 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>$(TargetFramework)</TargetFramework>
-    <Nullable>enable</Nullable>
-    <LangVersion>$(LangVersion)</LangVersion>
-    <VersionPrefix>$(BotSharpVersion)</VersionPrefix>
-    <GeneratePackageOnBuild>$(GeneratePackageOnBuild)</GeneratePackageOnBuild>
-    <GenerateDocumentationFile>$(GenerateDocumentationFile)</GenerateDocumentationFile>
-    <OutputPath>$(SolutionDir)packages</OutputPath>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="data\agents\6745151e-6d46-4a02-8de4-1c4f21c7da95\functions\handle_email_reader.json" />
-    <None Remove="data\agents\6745151e-6d46-4a02-8de4-1c4f21c7da95\templates\handle_email_reader.fn.liquid" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="data\agents\6745151e-6d46-4a02-8de4-1c4f21c7da95\functions\handle_email_reader.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-    <Content Include="data\agents\6745151e-6d46-4a02-8de4-1c4f21c7da95\templates\handle_email_reader.fn.liquid">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="MailKit" Version="4.7.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\..\onebrain\BusinessCore\BusinessCore.csproj" />
-    <ProjectReference Include="..\..\Infrastructure\BotSharp.Abstraction\BotSharp.Abstraction.csproj" />
-    <ProjectReference Include="..\..\Infrastructure\BotSharp.Core\BotSharp.Core.csproj" />
-  </ItemGroup>
-
-</Project>
