--- src/Plugins/BotSharp.Plugin.EmailReader/Enums/UtilityName.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BotSharp.Plugin.EmailReader.Enums;
-
-public class UtilityName
-{
-    public const string EmailReader = "email-reader";
-}
