--- src/Plugins/BotSharp.Plugin.EmailReader/Using.cs
@@ -1,18 +1,0 @@
-global using System;
-global using System.Collections.Generic;
-global using System.Text;
-global using BotSharp.Abstraction.Conversations;
-global using BotSharp.Abstraction.Plugins;
-global using System.Text.Json;
-global using BotSharp.Abstraction.Conversations.Models;
-global using System.Threading.Tasks;
-global using BotSharp.Abstraction.Functions;
-global using BotSharp.Abstraction.Agents.Models;
-global using BotSharp.Abstraction.Templating;
-global using Microsoft.Extensions.DependencyInjection;
-global using System.Linq;
-global using BotSharp.Abstraction.Utilities;
-global using BotSharp.Abstraction.Messaging;
-global using BotSharp.Abstraction.Messaging.Models.RichContent;
-global using BotSharp.Abstraction.Options;
-global using BotSharp.Abstraction.Messaging.Enums;
